// src/lib/api/models/EventStatistics.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
 
export type EventStatistics = {
    totalEvents?: number;
    publishedEvents?: number;
    draftEvents?: number;
    cancelledEvents?: number;
    completedEvents?: number;
    totalRegistrations?: number;
    totalAttendees?: number;
    totalRevenue?: number;
    pendingRevenue?: number;
};